You'll never know when you're done

+++ Проф.(1)Программа.py
@@ -1,11 +1,12 @@
 # Python 3.7.2
+# Calculating cosine of a powered by 2
 
 from math import degrees, cos
 
 # Entering number of degrees
 while True:
     try:
-        a = int(input('Enter integer representing degrees: '))
+        a = float(input('Enter integer representing degrees: '))
         break
 
     except ValueError:
